requirements.txt
~atplotlib==3.1.1
opencv_python==4.5.1.48
pytesseract==0.3.7
imageai==2.1.6
easyocr==1.3.0.1
requests==2.22.0
Keras==2.4.3
imutils==0.5.4
streamlit==0.79.0
streamlit_theme==0.58.0
numpy==1.18.5
Image==1.5.33
ipython==7.22.0
matplotlib==3.4.1
Pillow==8.2.0
scikit_learn==0.24.1
